server-kit/app/requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
grpcio>=1.66.0,<2.0.0
grpcio-tools>=1.66.0,<2.0.0
protobuf>=6.33.0,<7.0.0
pytest>=8.3.0,<9.0.0
httpx>=0.28.0,<1.0.0
PyYAML>=6.0.2,<7.0.0
